Re-set CardForm on CardScanningFragment
CardScanningFragment can be destroyed when the device has low resources.
Starting up CardIO is a great example of a low resource case. When this
happens the CardScanningFragment loses a reference to the CardForm. To
resolve it, the CardForm has to look for the fragment and re-set the
reference.

We also needed a mechanism that would prevent the CardScanningFragment
from re-creating CardIO on every onCreate call, especially when
re-creating.

Fixes #38